fix: 모달 창 열린 상태에서 뒤로가기 버튼 클릭 시, 모달이 닫히며 페이지 이동. 해당 로직 이벤트는 모달이 있을 때만 작동하도록 함
