+++ src/components/header/UserProfileButton.module.css
@@ -1,24 +1,3 @@
-.modal {
-  max-width: 90%;
-  min-height: 90%;
-
-  align-self: center;
-  align-items: center;
-  justify-content: center;
-  background-color: rgb(238, 238, 238);
-  display: flex;
-}
-
-.modal > button {
-  scale: 0.5;
-  position: absolute;
-  background-color: transparent;
-}
-
-.modal > button > * {
-  scale: 3.3;
-}
-
 .profile {
   border-radius: 50%;
   width: 30px;
